# alts: increase write record size max to 1MB