@arjan-bal 2025-08-14
Hi @kevinGC, I had a discussion with @dfawley about changing the buffer sizes earlier this year. Presently, the buffers used by ALTS can only grow, but not shrink. If there are hundreds of directpath channels (see internal doc [go/gcs-dp-connections-number-issue](http://goto.google.com/gcs-dp-connections-number-issue)) created in an idle client, these buffers can hold on to significant amount of memory if they are 1MB each. 

I [tried the scatter gather style of Java](https://github.com/grpc/grpc-go/compare/master...arjan-bal:alts-scatter-gather?expand=1), but we need to materialize the read data for decryption and that causes a copy operation. The performance was worse than the present implementation. A simpler and possibly equally performant solution could be to check if the `len(currentBuffer) > 32KB`  (initial size of the buffer) and the `nextFrameLen < 0.5 len(currentBuffer)`. If both are true copy the data into a new buffer from the pool of half the size. I decided not to pursue this further since the GCS benchmarks were showing that ALTS records were always around 16KB and there were diminishing returns on having very large buffer sizes.