@matthewstevenson88 2025-08-16
Update: The max frame size is not being set [here](https://github.com/grpc/grpc-go/blob/0ebea3ebca8720be615d2c2e270a2c55ebd0b818/credentials/alts/internal/handshaker/handshaker.go#L189), which means the ALTS max frame size negotiation is not happening during the handshake and (I think) means that the server is defaulting to 16KB frame size to be on the safe side. I think we should try setting it to a larger value (C++ has it set at 1024*1024), which should trigger the server to start sending large frames, and the other changes in this PR should (hopefully) lead the client to send large frames.

Could we run a quick test and see if setting the max frame size does the trick when talking to a real GCS server?